Add files via upload
Here I have worked on the movies to help a producer to make a movie which can break into top 250 in imdb list. My conclusion are that the movie will require at least 8/10 rating. The producer should work a movie with Drama as genre as the top 250 movies have 177 movies with genre as "Drama " and producer should hire Christopher Nolan as the director because he is the director seven such movies which made to top 250.
I came across #spilled issue while converting a column into a table because i was using a dynamic array formula such as unique and countif . I used
 =IFERROR(SUM(--(ISNUMBER(SEARCH(U2, L3:N252)))), 0) instead of countif in the table and
 =IFERROR(INDEX($AB$2:$AB$100, MATCH(0, COUNTIF($AC$1:AC1, $AB$2:$AB$100) + (COUNTIF($AB$2:$AB$100, "<>" & "") = 0), 0)), "") to overcome Unique issue. 
My version of excel did not have transpose function so i used =INDEX($L$2:$N$251, CEILING(ROWS($P$2:P2)/3, 1), MOD(ROWS($P$2:P2)-1, 3)+1) to Transpose the list.